before 2.2.20-beta2

+++ composeApp/src/commonMain/kotlin/ir/khebrati/audiosense/domain/useCase/spl/dbSpl.kt
@@ -30,6 +30,6 @@ private val normalEarHearingThresholds =
     )
 
 val Number.fromDbSpl: Int
-    get() = 20 * log10(this.toFloat()).toInt()
+    get() = 20.toInt() * (log10(this.toFloat())).toInt()
 val Number.dbSpl: Float
     get() = (10.0f.pow(this.toFloat() / 20f))
